+++ src/hooks/use-fetch-wpcom-sites/types/sync-support.ts
@@ -1,0 +1,7 @@
+export type SyncSupport =
+	| 'unsupported'
+	| 'syncable'
+	| 'needs-transfer'
+	| 'already-connected'
+	| 'jetpack-site'
+	| 'deleted';
