+++ src/preload.ts
@@ -4,10 +4,10 @@
 import '@sentry/electron/preload';
 import { SaveDialogOptions, contextBridge, ipcRenderer } from 'electron';
 import { LocaleData } from '@wordpress/i18n';
-import { SyncSite } from './hooks/use-fetch-wpcom-sites';
 import { ExportOptions } from './lib/import-export/export/types';
 import { BackupArchiveInfo } from './lib/import-export/import/types';
 import { promptWindowsSpeedUpSites } from './lib/windows-helpers';
+import type { SyncSite } from './hooks/use-fetch-wpcom-sites/types/sync-site';
 import type { LogLevel } from './logging';
 
 const api: IpcApi = {
